+++ src/main/java/org/mathevaluator/api/FunctionController.java
@@ -20,24 +20,24 @@ public class FunctionController {
 
     @RequestMapping(value = "", method = RequestMethod.GET, produces = "application/json")
     public String function(@RequestParam Map<String, String> params) {
-	JSONObject json = new JSONObject();
-	try {
-	    Evaluator evaluator = new Evaluator(params.get("f"));
-	    Map<String, Expression> variables = new HashMap<>();
-	    for (Entry<String, String> entry : params.entrySet()) {
+        JSONObject json = new JSONObject();
+        try {
+            Evaluator evaluator = new Evaluator(params.get("f"));
+            Map<String, Expression> variables = new HashMap<>();
+            for (Entry<String, String> entry : params.entrySet()) {
 
-		if(entry.getKey().equals("f") || entry.getValue().isEmpty()) {
-		    continue;
-		}
-		variables.put(entry.getKey(), new NumberExpression(Double.valueOf(entry.getValue())));
-	    }
+                if(entry.getKey().equals("f") || entry.getValue().isEmpty()) {
+                    continue;
+                }
+                variables.put(entry.getKey(), new NumberExpression(Double.valueOf(entry.getValue())));
+            }
 
-	    json.append("result", evaluator.interpret(variables));
-	} catch (JSONException | InvalidExpressionException e) {
-	    json.append("error", e.getMessage());
-	    e.printStackTrace();
-	}
-	return json.toString();
+            json.append("result", evaluator.interpret(variables));
+        } catch (JSONException | InvalidExpressionException e) {
+            json.append("error", e.getMessage());
+            e.printStackTrace();
+        }
+        return json.toString();
     }
 
 }