+++ src/main/java/org/mathevaluator/interpreter/InvalidExpressionException.java
@@ -2,10 +2,10 @@ package org.mathevaluator.interpreter;
 
 public class InvalidExpressionException extends Exception {
 
-	private static final long serialVersionUID = 616247988149970779L;
+    private static final long serialVersionUID = 616247988149970779L;
 
-	public InvalidExpressionException(String expression) {
-		super(expression);
-	}
+    public InvalidExpressionException(String expression) {
+	super(expression);
+    }
 
 }